refactor: config.pyのモジュール分割（1100行→385行）
- 設定クラスを論理的にグループ化してsettings/ディレクトリに分割
  - api_settings.py: API設定（APIConfig）
  - weather_settings.py: 天気関連設定（WeatherConfig, WeatherConstants）
  - app_settings.py: アプリ設定（AppSettings, ServerConfig, SystemConstants）
  - llm_settings.py: LLM設定（LLMConfig, LangGraphConfig）
  - comment_settings.py: コメント設定（CommentConfig, SevereWeatherConfig）
  - ui_data_settings.py: UI・データ設定（UISettings, GenerationSettings, DataSettings）

- config.pyを大幅に簡潔化（385行）
  - 各設定モジュールからインポート
  - Configクラスで統合管理
  - ヘルパー関数は維持

- app_config.pyの型エラーを修正
  - generation_timeout → timeout
  - 不要なisinstance呼び出しを削除
